refactor: split FeesUpdater
